+++ examples/src/virtual/react/hook/window-scroll/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @af-utils/virtual-examples-react-hook-window-scroll
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @af-utils/virtual-react@0.0.25
+
 ## 0.0.4
 
 ### Patch Changes